Commit Four - Map Functionality
In this commit I have implemented a simple map, using googlemaps and
javascript code, to indicate the location of Manchester United's home
stadium Old Trafford. However, for some reason, the map only renders
once you've inspected the element of the page (Firefox). I am not sure
if it is because it has been placed within <section> tags. Will
investigate further...