Commit Six - Styling
I have began to change the presentation of my website, trying to make it
look good. A few things I've changed is the addition of the background
colour and the font colour change. Additional changes will come later
on. Next step is to complete the content of the website, i.e the rest of
the news, team and fixtures