Commit Seven - Fixtures and Results
I have now added content to be displayed in a table. I have added a lot
of styles to this table and pictures as well. I will next add a
javascript function to the content of the table, namely the SCORE
column. This javascript should allow users to change the timezone of
future matches so that they can see when the match will play in their
local time zone.