Commit Seven - Javascript TimeChanger and Modals
This has been a long one to try and implement. I managed to get a pop up
modal to appear and display the time zone functions,  and I managed to
get the time zone to change with javascrip as well. See time.js.
However, I cannot get it to work with other modals, only one modal seems
to work.