Commit Eight - Rest of the content added
In this commit I have added the rest of my website content, minus the
last few pictures. I had to revert to a backup index file as the main
one got too many errors when working on, therefore indextest.html is the
main file now (To be renamed later). I fixed the navigation issues as
well by using javascript functions. These can be found at the top of the
file.